fix(config): refine particle display settings for enhanced performance
